虚拟币获取模块-参加竞赛获取完成10.19-2

+++ src/main/java/com/example/xunibibackend/service/GainService.java
@@ -4,7 +4,7 @@ import com.example.xunibibackend.entity.Achievement;
 
 public interface GainService {
 
-    boolean checkIn(Long id);
+    boolean checkIn(Integer id);
 
     boolean rewardAchievement(Achievement achievement);
